requirements.txt
snowflake-connector-python
altair==4.0
streamlit
datetime
pandas